+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - `etro.layer.BaseAudio` (which all audio layers subclass).
+- `etro.layer.Oscillator`.
 - `etro.effect.Audio` (which all audio effects subclass).
 - `etro.effect.Volume`.
 - `etro.effect.Panner`.